+++ lib/page/hello/recom/recom_user_page.dart
@@ -13,13 +13,12 @@
  *  this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import 'dart:io';
 
-import 'package:flutter/material.dart';
-import 'package:flutter_mobx/flutter_mobx.dart';
-import 'package:pixez/component/painer_card.dart';
-import 'package:pixez/i18n.dart';
+import 'package:flutter/widgets.dart';
+import 'package:pixez/constants.dart';
 import 'package:pixez/page/hello/recom/recom_user_store.dart';
+import 'package:pixez/page/hello/recom/user_page/fluent_state.dart';
+import 'package:pixez/page/hello/recom/user_page/material_state.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 
 class RecomUserPage extends StatefulWidget {
@@ -28,82 +27,34 @@ class RecomUserPage extends StatefulWidget {
   const RecomUserPage({Key? key, this.recomUserStore}) : super(key: key);
 
   @override
-  _RecomUserPageState createState() => _RecomUserPageState();
+  RecomUserPageStateBase createState() {
+    if (Constants.isFluentUI)
+      return FluentRecomUserPageState();
+    else
+      return MaterialRecomUserPageState();
+  }
 }
 
-class _RecomUserPageState extends State<RecomUserPage> {
-  late RefreshController _refreshController;
-  late RecomUserStore _recomUserStore;
+abstract class RecomUserPageStateBase extends State<RecomUserPage> {
+  late RefreshController refreshController;
+  late RecomUserStore recomUserStore;
 
   @override
   void initState() {
-    _refreshController =
+    refreshController =
         RefreshController(initialRefresh: widget.recomUserStore == null);
-    _recomUserStore =
-        widget.recomUserStore ?? RecomUserStore(controller: _refreshController);
+    recomUserStore =
+        widget.recomUserStore ?? RecomUserStore(controller: refreshController);
     if (widget.recomUserStore != null) {
-      _recomUserStore.controller = _refreshController;
+      recomUserStore.controller = refreshController;
     }
     super.initState();
   }
 
   @override
   void dispose() {
-    _recomUserStore.controller = null;
-    _refreshController.dispose();
+    recomUserStore.controller = null;
+    refreshController.dispose();
     super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Observer(builder: (context) {
-      return Scaffold(
-        appBar: AppBar(
-          title: Text(I18n.of(context).recommend_for_you),
-        ),
-        body: SmartRefresher(
-          header: Platform.isAndroid
-              ? MaterialClassicHeader(
-                  color: Theme.of(context).colorScheme.secondary,
-                )
-              : ClassicHeader(),
-          controller: _refreshController,
-          enablePullDown: true,
-          enablePullUp: true,
-          onRefresh: () => _recomUserStore.fetch(),
-          onLoading: () => _recomUserStore.next(),
-          footer: CustomFooter(
-            builder: (BuildContext context, LoadStatus? mode) {
-              Widget body;
-              if (mode == LoadStatus.idle) {
-                body = Text(I18n.of(context).pull_up_to_load_more);
-              } else if (mode == LoadStatus.loading) {
-                body = CircularProgressIndicator();
-              } else if (mode == LoadStatus.failed) {
-                body = Text(I18n.of(context).loading_failed_retry_message);
-              } else if (mode == LoadStatus.canLoading) {
-                body = Text(I18n.of(context).let_go_and_load_more);
-              } else {
-                body = Text(I18n.of(context).no_more_data);
-              }
-              return Container(
-                height: 55.0,
-                child: Center(child: body),
-              );
-            },
-          ),
-          child: _recomUserStore.users.isNotEmpty
-              ? ListView.builder(
-                  itemCount: _recomUserStore.users.length,
-                  itemBuilder: (context, index) {
-                    final data = _recomUserStore.users[index];
-                    return PainterCard(
-                      user: data,
-                    );
-                  })
-              : Container(),
-        ),
-      );
-    });
   }
 }
