+++ lib/page/hello/recom/user_page/material_state.dart
@@ -1,0 +1,78 @@
+/*
+ * Copyright (C) 2020. by perol_notsf, All rights reserved
+ *
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 3 of the License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful, but WITHOUT ANY
+ *  WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ *  FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with
+ *  this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:pixez/component/painer_card.dart';
+import 'package:pixez/i18n.dart';
+import 'package:pixez/page/hello/recom/recom_user_page.dart';
+import 'package:pull_to_refresh/pull_to_refresh.dart';
+
+class MaterialRecomUserPageState extends RecomUserPageStateBase {
+  @override
+  Widget build(BuildContext context) {
+    return Observer(builder: (context) {
+      return Scaffold(
+        appBar: AppBar(
+          title: Text(I18n.of(context).recommend_for_you),
+        ),
+        body: SmartRefresher(
+          header: Platform.isAndroid
+              ? MaterialClassicHeader(
+                  color: Theme.of(context).colorScheme.secondary,
+                )
+              : ClassicHeader(),
+          controller: refreshController,
+          enablePullDown: true,
+          enablePullUp: true,
+          onRefresh: () => recomUserStore.fetch(),
+          onLoading: () => recomUserStore.next(),
+          footer: CustomFooter(
+            builder: (BuildContext context, LoadStatus? mode) {
+              Widget body;
+              if (mode == LoadStatus.idle) {
+                body = Text(I18n.of(context).pull_up_to_load_more);
+              } else if (mode == LoadStatus.loading) {
+                body = CircularProgressIndicator();
+              } else if (mode == LoadStatus.failed) {
+                body = Text(I18n.of(context).loading_failed_retry_message);
+              } else if (mode == LoadStatus.canLoading) {
+                body = Text(I18n.of(context).let_go_and_load_more);
+              } else {
+                body = Text(I18n.of(context).no_more_data);
+              }
+              return Container(
+                height: 55.0,
+                child: Center(child: body),
+              );
+            },
+          ),
+          child: recomUserStore.users.isNotEmpty
+              ? ListView.builder(
+                  itemCount: recomUserStore.users.length,
+                  itemBuilder: (context, index) {
+                    final data = recomUserStore.users[index];
+                    return PainterCard(
+                      user: data,
+                    );
+                  })
+              : Container(),
+        ),
+      );
+    });
+  }
+}
